- header parser - deleted Meter class and changed to Fraction

+++ abcPlayer/src/datatypes/Fraction.java
@@ -24,7 +24,7 @@ public class Fraction {
 	}
 	
 	public double getValue() {
-		return numerator/denominator;
+		return ((double)numerator)/denominator;
 	}
 	
 	@Override
